refactor: etl package split into dto and service
